update route corrected

+++ backend/controllers/goalController.js
@@ -1,6 +1,7 @@
 const asyncHandler = require('express-async-handler')
 
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
 
 //@descr  Get goals
 //@route  GET api/goals
@@ -8,7 +9,7 @@ const Goal = require('../models/goalModel')
 
 
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find({user: req.user.id})
+    const goals = await Goal.find({ user: req.user.id })
 
     res.status(200).json(goals)
 })
@@ -19,7 +20,7 @@ const getGoals = asyncHandler(async (req, res) => {
 
 
 const setGoal = asyncHandler(async (req, res) => {
-    if(!req.body.text) {
+    if (!req.body.text) {
         res.status(400)
         throw new Error('Please add a text field')
     }
@@ -42,13 +43,23 @@ const updateGoal = asyncHandler(async (req, res) => {
 
     const goal = await Goal.findById(req.params.id)
 
-
-    if(!goal) {
+    if (!goal) {
         res.status(400)
         throw new Error('Goal not found')
     }
 
-    const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body.text, {new: true})
+    const user = await User.findById(req.user.id)
+    if (!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    if (goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorised')
+    }
+
+    const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, {text: req.body.text}, { new: true })
 
 
     res.status(200).json(updatedGoal)
@@ -63,14 +74,25 @@ const deleteGoal = asyncHandler(async (req, res) => {
 
     const goal = await Goal.findById(req.params.id)
 
-    if(!goal) {
+    if (!goal) {
         res.status(400)
         throw new Error('Goal not found')
     }
 
+    const user = await User.findById(req.user.id)
+    if (!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    if (goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorised')
+    }
+
     await goal.deleteOne()
 
-    res.status(200).json({id: req.params.id})
+    res.status(200).json({ id: req.params.id })
 })
